refactor: use same analysis functions in both personal and shared
