+++ client/utils/analysis.ts
@@ -1,0 +1,40 @@
+import {
+  MainCategoryKey,
+  subcategories,
+  SubcategoryKey,
+} from 'data/categories';
+import { SharedTxn } from 'pages/shared/trackers/[trackerId]';
+import { Transaction } from 'types/Transaction';
+
+export function calculateTotal(txns: Transaction[] | SharedTxn[]) {
+  let total = 0;
+  for (const txn of txns) {
+    total += txn.amount;
+  }
+  return total;
+}
+
+export function totalsByMainCategory(txns: Transaction[] | SharedTxn[]) {
+  const totals = {} as Record<MainCategoryKey, number>;
+  for (const txn of txns) {
+    const mainCategory = subcategories[txn.category].mainCategory;
+    if (!totals[mainCategory]) totals[mainCategory] = 0;
+    totals[mainCategory] += txn.amount;
+  }
+  return Object.entries(totals).map(([category, total]) => ({
+    category,
+    total,
+  }));
+}
+
+export function totalsBySubCategory(txns: Transaction[] | SharedTxn[]) {
+  const totals = {} as Record<SubcategoryKey, number>;
+  for (const txn of txns) {
+    if (!totals[txn.category]) totals[txn.category] = 0;
+    totals[txn.category] += txn.amount;
+  }
+  return Object.entries(totals).map(([category, total]) => ({
+    category,
+    total,
+  }));
+}
